Split all the sound stuff into a separate sound.cc file.
Did not add file to source/MSVC/crawl.vcxproj.filters; I assume this
auto-generated.
