+++ crawl-ref/source/Makefile.obj
@@ -202,6 +202,7 @@ show.o \
 showsymb.o \
 skill-menu.o \
 skills.o \
+sound.o \
 species.o \
 spl-book.o \
 spl-cast.o \
