+++ crawl-ref/source/android-project/jni/src/Android.mk
@@ -222,6 +222,7 @@ LOCAL_SRC_FILES := $(SDL_PATH)/src/main/android/SDL_android_main.c \
     $(CRAWL_PATH)/showsymb.cc \
     $(CRAWL_PATH)/skill-menu.cc \
     $(CRAWL_PATH)/skills.cc \
+    $(CRAWL_PATH)/sound.cc \
     $(CRAWL_PATH)/species.cc \
     $(CRAWL_PATH)/spl-book.cc \
     $(CRAWL_PATH)/spl-cast.cc \
