+++ crawl-ref/source/l-crawl.cc
@@ -33,6 +33,7 @@ module "crawl"
 #include "perlin.h"
 #include "prompt.h"
 #include "religion.h"
+#include "sound.h"
 #include "state.h"
 #include "state.h"
 #include "stringutil.h"
